Revert "Hotfixes"
