+++ api/src/main/java/com/chinita/company/gestionprocedimientos/api/Controlador/PersonaControlador.java
@@ -1,0 +1,17 @@
+package com.chinita.company.gestionprocedimientos.api.Controlador;
+
+
+import com.chinita.company.gestionprocedimientos.api.Servicio.PersonaServicio;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping ("/personas")
+public class PersonaControlador {
+
+    private PersonaServicio personaServicio;
+
+    public PersonaControlador(PersonaServicio personaServicio) {
+        this.personaServicio = personaServicio;
+    }
+}
