+++ api/src/main/java/com/chinita/company/gestionprocedimientos/api/Repository/PersonaRepository.java
@@ -1,0 +1,7 @@
+package com.chinita.company.gestionprocedimientos.api.Repository;
+
+import com.chinita.company.gestionprocedimientos.api.Modelo.Persona;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PersonaRepository extends JpaRepository<Persona, Integer> {
+}
