+++ api/src/main/java/com/chinita/company/gestionprocedimientos/api/Servicio/PersonaServicio.java
@@ -1,0 +1,9 @@
+package com.chinita.company.gestionprocedimientos.api.Servicio;
+
+import com.chinita.company.gestionprocedimientos.api.Modelo.Persona;
+
+import java.util.List;
+
+public interface PersonaServicio {
+
+}
