+++ api/src/main/java/com/chinita/company/gestionprocedimientos/api/Servicio/PersonaServicioImplementacion.java
@@ -1,0 +1,14 @@
+package com.chinita.company.gestionprocedimientos.api.Servicio;
+
+import com.chinita.company.gestionprocedimientos.api.Repository.PersonaRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class PersonaServicioImplementacion implements PersonaServicio{
+
+    private PersonaRepository personaRepository;
+
+    public PersonaServicioImplementacion(PersonaRepository personaRepository) {
+        this.personaRepository = personaRepository;
+    }
+}
