--- api/src/main/java/com/chinita/company/gestionprocedimientos/api/repository/PersonaRepository.java
@@ -1,9 +1,0 @@
-package com.chinita.company.gestionprocedimientos.api.repository;
-
-import com.chinita.company.gestionprocedimientos.api.modelo.Persona;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PersonaRepository extends JpaRepository<Persona, Integer> {
-}
