--- api/src/main/java/com/chinita/company/gestionprocedimientos/api/servicio/PersonaServicio.java
@@ -1,11 +1,0 @@
-package com.chinita.company.gestionprocedimientos.api.servicio;
-
-import com.chinita.company.gestionprocedimientos.api.modelo.Persona;
-import java.util.List;
-
-public interface PersonaServicio {
-
-    List<Persona> getPersonas();
-    Persona addPersona(Persona persona);
-
-}
