--- api/src/main/resources/application-dev.yml
@@ -1,31 +1,0 @@
-spring:
-  jpa:
-    show-sql: true
-    hibernate:
-      default_schema: public
-      ddl-auto: update
-    properties:
-      hibernate:
-        order_inserts: true
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-        jdbc:
-          batch_size: 50
-          lob:
-            non_contextual_creation: true
-        temp:
-          use_jdbc_metadata_defaults: false
-
-  datasource:
-    type: com.zaxxer.hikari.HikariDataSource
-    url: jdbc:postgresql://localhost:5432/procedimientos-management
-    driver-class-name: org.postgresql.Driver
-    username:
-    password:
-    hikari:
-      idle-timeout: 10000
-      maximum-pool-size: 5
-
-logging:
-  level:
-    root: INFO
-    org.springframework: ERROR
