--- api/src/main/resources/application-heroku.yml
@@ -1,31 +1,0 @@
-spring:
-  jpa:
-    show-sql: true
-    hibernate:
-      default_schema: public
-      ddl-auto: update
-    properties:
-      hibernate:
-        order_inserts: true
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-        jdbc:
-          batch_size: 50
-          lob:
-            non_contextual_creation: false
-        temp:
-          use_jdbc_metadata_defaults: false
-
-  datasource:
-    type: com.zaxxer.hikari.HikariDataSource
-    url: jdbc:postgresql://ec2-34-197-212-240.compute-1.amazonaws.com:5432/d3j2fmkf7rv3jn?ssl=true&sslfactory=org.postgresql.ssl.NonValidatingFactory&sslmode=require
-    driver-class-name: org.postgresql.Driver
-    username: vzcrvgkdqnxgzl
-    password: 1c87ee2ef56e92a1ef21a44c3a17e0490017ea2eef77fea2898ad004e7cda64a
-    hikari:
-      idle-timeout: 10000
-      maximum-pool-size: 5
-
-logging:
-  level:
-    root: INFO
-    org.springframework: ERROR
