--- api/src/main/resources/application.yml
@@ -1,15 +1,0 @@
-# Server properties
-server:
-  servlet:
-    context-path: /gestion-procedimientos
-  port: 8080
-
-# Spring properties
-spring:
-  # Dev is active by default. This value can be overwritten by java arg param
-  profiles:
-    active: heroku
-  application:
-    name: gestion-procedimientos
-  messages:
-    basename: bundles/apiLanguageBundle
