--- api/src/main/resources/bundles/apiLanguageBundle.properties
@@ -1,9 +1,0 @@
-# This is an UTF-8 File. Please take care of itdetalle_requerido
-# Accents: a = \u00e1, e = \u00e9, i = \u00ed, o = \u00f3, u = \u00fa,  ?= \u00f1
-# Accents Mayus: A = \u00C1, E = \u00C9, I = \u00CD, O = \u00D3, U = \u00DA,  = \u00D1
-# New line = \u000a
-#------------------------------------------------API-----------------------------------------------------------------
-
-exception.data_duplicate_name.proyecto=la persona con ese documento ya existe
-exception.data_id_not_exist.proyecto=la persona con ese documento no existe
-exception.data_not_found.proyecto=Personas no encontrada
