--- web/src/app/Persona/persona-list/persona-list.component.html
@@ -1,1 +1,0 @@
-<p>persona-list works!</p>
