--- web/src/app/Persona/persona-list/persona-list.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-persona-list',
-  templateUrl: './persona-list.component.html',
-  styleUrls: ['./persona-list.component.css']
-})
-export class PersonaListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
