--- web/src/app/app.module.ts
@@ -1,18 +1,0 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-
-import { AppComponent } from './app.component';
-import { PersonaListComponent } from './Persona/persona-list/persona-list.component';
-
-@NgModule({
-  declarations: [
-    AppComponent,
-    PersonaListComponent
-  ],
-  imports: [
-    BrowserModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
-export class AppModule { }
